refactor on etls
