+++ etls/scripts/base_etl_class.py
@@ -1,0 +1,8 @@
+import abc
+
+class ETL(abc.ABC):
+
+
+    @abc.abstractmethod
+    def __call__(self):
+        pass
